requirements.txt
click==7.1.2
fastapi==0.62.0
h11==0.9.0
httptools==0.1.1
pydantic==1.7.3
email_validator
requests==2.22.0
starlette==0.13.6
SQLAlchemy==1.4
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
black==20.8b1
flake8==3.9.0
fastapi_utils==0.2.1
pytest
psycopg2-binary
python-jose
passlib
bcrypt
python-multipart